refactored individual methods into separate files
