+++ src/core/firstOrDefault.js
@@ -1,0 +1,23 @@
+var Enumerable = require('../enumerable');
+
+Enumerable.fn.firstOrDefault = function(fn)
+{
+    if(this.collection.length > 0)
+    {
+        return this.collection[0];
+    }
+
+    if(fn !== null && fn !== undefined)
+    {
+        if(typeof fn === "function")
+        {
+            return fn();
+        }
+        else
+        {
+            return fn;
+        }
+    }
+
+    return null;
+};
