+++ src/core/skipWhile.js
@@ -1,0 +1,24 @@
+var Enumerable = require('../enumerable');
+
+Enumerable.fn.skipWhile = function(fn)
+{
+    var results = [];
+
+    var canSkip = (fn !== null && fn !== undefined && typeof fn === "function");
+
+    if(canSkip)
+    {
+        for(var i = 0; i < this.collection.length; i++)
+        {
+            var item = this.collection[i];
+            var _callbackResult = fn(i, item, this.collection);
+
+            if(_callbackResult !== true)
+            {
+                results.push(this.collection[i]);
+            }
+        }
+    }
+
+    return results.asEnumerable();
+};
