+++ src/core/toArray.js
@@ -1,0 +1,6 @@
+var Enumerable = require('../enumerable');
+
+Enumerable.fn.toArray = function()
+{
+    return this.collection;
+};
